+++ programs/switchboard-on-demand/src/lib.rs
@@ -4,11 +4,22 @@ use anchor_lang::program;
 use anchor_lang::AnchorDeserialize;
 use solana_program::pubkey::Pubkey;
 
+#[cfg(feature = "mainnet-beta")]
 declare_id!("SBondMDrcV3K4kxZR1HNVT7osZxAHVHgYXL5Ze1oMUv");
+#[cfg(not(feature = "mainnet-beta"))]
+declare_id!("Aio4gaXjXzJNVLtzwtNVmSqGKpANtXhybbkhtAC94ji2");
 
 #[program]
 pub mod switchboard_on_demand {}
 pub const SB_ON_DEMAND_PRECISION: u32 = 18;
+
+#[repr(C)]
+#[derive(Clone, Copy, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct CompactResult {
+    pub std_dev: f32,
+    pub mean: f32,
+    pub slot: u64,
+}
 
 #[repr(C)]
 #[derive(Clone, Copy, Debug, bytemuck::Pod, bytemuck::Zeroable)]
@@ -25,7 +36,8 @@ pub struct CurrentResult {
     pub min_value: i128,
     /// The maximum value of the submissions needed for quorom size
     pub max_value: i128,
-    pub padding1: [u8; 8],
+    pub num_samples: u8,
+    pub padding1: [u8; 7],
     /// The slot at which this value was signed.
     pub slot: u64,
     /// The slot at which the first considered submission was made
@@ -147,17 +159,19 @@ pub struct PullFeedAccountData {
     pub max_variance: u64,
     pub min_responses: u32,
     pub name: [u8; 32],
-    _padding1: [u8; 3],
-    pub sample_size: u8,
+    _padding1: [u8; 2],
+    pub historical_result_idx: u8,
+    pub min_sample_size: u8,
     pub last_update_timestamp: i64,
     pub lut_slot: u64,
-    pub ipfs_hash: [u8; 32], // deprecated
+    _reserved1: [u8; 32], // deprecated
     pub result: CurrentResult,
     pub max_staleness: u32,
-    _ebuf4: [u8; 20],
+    _padding2: [u8; 12],
+    pub historical_results: [CompactResult; 32],
+    _ebuf4: [u8; 8],
     _ebuf3: [u8; 24],
     _ebuf2: [u8; 256],
-    _ebuf1: [u8; 512],
 }
 
 impl PullFeedAccountData {