+++ sdk/src/constants/perpMarkets.ts
@@ -874,6 +874,28 @@ export const MainnetPerpMarkets: PerpMarketConfig[] = [
 		pythFeedId:
 			'0x62742a997d01f7524f791fdb2dd43aaf0e567d765ebf8fd0406a994239e874d4',
 	},
+	{
+		fullName: 'MOODENG',
+		category: ['Solana', 'Meme'],
+		symbol: 'MOODENG-PERP',
+		baseAssetSymbol: 'MOODENG',
+		marketIndex: 45,
+		oracle: new PublicKey('21gjgEcuDppthwV16J1QpFzje3vmgMp2uSzh7pJsG7ob'),
+		launchTs: 1727965864000,
+		oracleSource: OracleSource.PYTH_PULL,
+		pythFeedId:
+			'0xffff73128917a90950cd0473fd2551d7cd274fd5a6cc45641881bbcc6ee73417',
+	},
+	{
+		fullName: 'WARWICK-FIGHT-WIN-BET',
+		category: ['Prediction', 'Sport'],
+		symbol: 'WARWICK-FIGHT-WIN-BET',
+		baseAssetSymbol: 'WARWICK-FIGHT-WIN',
+		marketIndex: 46,
+		oracle: new PublicKey('Dz5Nvxo1hv7Zfyu11hy8e97twLMRKk6heTWCDGXytj7N'),
+		launchTs: 1727965864000,
+		oracleSource: OracleSource.Prelaunch,
+	},
 ];
 
 export const PerpMarkets: { [key in DriftEnv]: PerpMarketConfig[] } = {